+++ internal/dashboard/business/oceanbase/obcluster_log.go
@@ -1,4 +1,4 @@
-package obcluster
+package oceanbase
 
 import (
 	"context"
